pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tia-service-broadcaster"
version = "1.0.0"
description = "Signal Broadcasting Service for Mudrex Trade Ideas Automation"
authors = [{name = "DecentralizedJM"}]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "python-telegram-bot==21.0.1",
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "websockets==14.1",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.1",
    "aiosqlite==0.20.0",
    "databases==0.9.0",
    "cryptography==44.0.0",
    "python-dateutil==2.9.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
    "httpx==0.28.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
